Post Like table and related create and init operations are handled.
** There is used dummy data because there aren't any user and post
tables created yet!

- Create and initialize of the post like table is implemented.
- In home page the like button of post is implemented and prepared for
the click. It has also the feature that allows just one click per user.
- In my gokturk.html, creation of table is implemented but it will be
completed properly after the related tables are implemented in project.
- within the handler and service python file, proper operations and its
directions are implemented at first.
- In server.py like button which has connection to the like of posts are
implemented.
- Proper changes are done in template files to prepare the like button
to be clicked and counted by doing this.
- The number of count of likes as it is shown on posts are updated only
when the page is refreshed ( It is under construction - demo version).