Delete op. for pLikes table is implemented and like button is updated.
- Select operation has already added before counting the like numbers.
- Delete operation database queries and postgresql connection is
implemented.
- Error handling and related operation is implemented in handler.
- Still the dummy data is used to demonstrate the delete operation
because the pLikes table has multiple relations with the User and Post
tables which are not created yet.
-- In home page, the responsive Like button for the posts are redesigned
and implemented as the one like per the post.(Also the like per user
features will be implemented when relations are fully created and
implemented.)
 - When the first click on Like button increase the counter as 1 from 0,
this operation is displayed in pLikes table which means it is added;
when it is shown as 1, then click again it becomes 0 which means that
the like is retrieved so it is deleted from pLikes table.
_