Add,Delete,Update, Select operations for Comment Table are created.
- Comment Table is connected with the Post Table via implementing delete
and update operations as cascade which means if a post is deleted, then
all related comments are removed either.

- Add operation has no extra page, it is implemented within the related
posts. It is because comments are connected with the posts. Summernote
which is open sourced - java script based text editor is implemented
under the posts and their comments as a comment.

- Delete operation is performed by clicking red trash-box button under
the comments. When it is deleted, related post should be still there.
But if the user deletes a post, then all comments related to this post
are deleted either. There are no extra page for delete operation because
of its post-related situation.

- Update operation is implemented under the comment section. Clicking
the yellow-pencil button, comment which is already connected its post
can be modified. There is an extra page for updating a comment as an
exception.

- Add, delete, update and select operation queries are implemented in
comment_service and the whole path is implemented in the handler python
file.